refactor router and main
